// types/country.ts
export type Country = {
  name: {
    common: string;
    official: string;
    nativeName: { [key: string]: { official: string; common: string } };
  };
  cca3: string;
  currencies: { [key: string]: { name: string; symbol: string } };
  capital: string[];
  region: string;
  subregion: string;
  languages: { [key: string]: string };
  population: number;
  flags: {
    png?: string;
    svg?: string;
    alt?: string;
  };
  borders: string[];
};